+++ bin/app.php
@@ -5,4 +5,4 @@ declare(strict_types=1);
 use BEAR\Skeleton\Bootstrap;
 
 require dirname(__DIR__) . '/autoload.php';
-exit((new Bootstrap())(PHP_SAPI === 'cli' ? 'cli-hal-api-app' : 'hal-api-app', $GLOBALS, $_SERVER, getenv('APP_WRITE_DIR') ?: null));
+exit((new Bootstrap())(PHP_SAPI === 'cli' ? 'cli-hal-api-app' : 'hal-api-app', $GLOBALS, $_SERVER));